requirements.txt
-i https://pypi.org/simple
-e .
asn1crypto==0.24.0
cffi==1.12.3
click==7.0
cryptography==2.6.1
cycler==0.10.0
flask-cors==3.0.7
flask==1.0.3
future==0.17.1
itsdangerous==1.1.0
jinja2==2.10.1
kiwisolver==1.1.0
markupsafe==1.1.1
matplotlib==3.1.0
numpy==1.16.3
onnx==1.5.0
pillow==6.0.0
protobuf==3.7.1
pycparser==2.19
pyopenssl==19.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
six==1.12.0
torch==1.1.0
torchsummary==1.5.1
torchvision==0.2.2.post3
tqdm==4.32.1
typing-extensions==3.7.2
typing==3.6.6
werkzeug==0.15.4
tb-nightly==1.14.0a20190522
